+++ packages/flutter/test/native/sentry_native_java_test_web.dart
@@ -1,5 +1,5 @@
 void main() {
   // This test file exists only to satisfy the compiler when running web tests.
-  // The actual tests in android_envelope_sender_test_real.dart are only
+  // The actual tests in sentry_native_java_test_real.dart are only
   // executed on VM platforms.
 }
